UPDATES: - changes to fare calculations, additional column in data file. Fare doesn't count to payment expanses in up table at all and is already added to salary in local table. Fare cost must be subtracted from salary. - new file pracov.csv contains start/end of employment, type and start of pension. Not implemented yet, just added to project. - changes is table structure - removing redundant files like structure.py, etc. from project - code organization
TOOD:
- delete data from later columns in up table list 3
- delete cost in list 3 if is 13 600 in up table
- add cost formula 13 600 - x to list 3 in up table
- make local copy of structure.json and use that instead of rewriting global
- use csv headers as dictionary keys for better code readability
- add start/end employment, type and start of pension to new employees
- open src folder and output folder for manual copying of checked files

+++ functions.py
@@ -4,6 +4,7 @@ import tkinter as tk
 import subprocess
 from tkinter import simpledialog
 from tkinter.filedialog import askopenfilename, askdirectory
+from tkinter.messagebox import askyesno
 from os import remove
 
 
@@ -123,44 +124,51 @@ def save_ins(c_name, my_tree):
 def rename_tab(window, companies, last_data, tab, tabs):
     c_name = tabs.tab(tab)['text']
     btn = window.nametowidget(tab + '.!checkbutton')
-    new_name = simpledialog.askstring("Input", "Test", parent=window)
+    new_name = simpledialog.askstring("Prejmenovat", "Novy nazev", parent=window)
+    
+    confirm = askyesno(title="POZOR", message="Opravdu chcete prejmenovat firmu?")
+    
+    if confirm:
 
-    if c_name in companies:
-        index = companies.index(c_name)
-        companies.pop(index)
-        companies.append(new_name)
+        if c_name in companies:
+            index = companies.index(c_name)
+            companies.pop(index)
+            companies.append(new_name)
 
-    tabs.tab(tab, text=new_name)
-    new_data = {new_name if k == c_name else k:v for k, v in last_data.items()}
-    btn['text'] = new_name
+        tabs.tab(tab, text=new_name)
+        new_data = {new_name if k == c_name else k:v for k, v in last_data.items()}
+        btn['text'] = new_name
 
-    with open(f'structure.json', 'w') as outfile:
-        outfile.write(json.dumps(new_data, indent=4))
+        with open(f'structure.json', 'w') as outfile:
+            outfile.write(json.dumps(new_data, indent=4))
 
-    if Path(f'insurance_codes_{c_name}.json').exists():
-        Path(f'insurance_codes_{c_name}.json').rename(f'insurance_codes_{new_name}.json')
+        if Path(f'insurance_codes_{c_name}.json').exists():
+            Path(f'insurance_codes_{c_name}.json').rename(f'insurance_codes_{new_name}.json')
 
 
 def delete_tab(last_data, companies, tabs):
-    c_name = tabs.tab(tabs.select())['text']
-    ins_file = Path(f'insurance_codes_{c_name}.json')
 
-    if c_name in last_data:
-        del last_data[c_name]
-        with open(f'structure.json', 'w') as outfile:
-            outfile.write(json.dumps(last_data, indent=4))
+    confirm = askyesno(title="POZOR", message="Chcete smazat firmu?")
+    
+    if confirm:
+        c_name = tabs.tab(tabs.select())['text']
+        ins_file = Path(f'insurance_codes_{c_name}.json')
 
-    if c_name in companies:
-        companies.remove(c_name)
+        if c_name in last_data:
+            del last_data[c_name]
+            with open(f'structure.json', 'w') as outfile:
+                outfile.write(json.dumps(last_data, indent=4))
 
-    if ins_file.exists():
-        print('deleting insurance file.')
-        remove(ins_file)
-    else:
-        print('No insurance file found.')
+        if c_name in companies:
+            companies.remove(c_name)
 
-    tabs.forget(tabs.select())
-    print(tabs.children)
+        if ins_file.exists():
+            print('deleting insurance file.')
+            remove(ins_file)
+        else:
+            print('No insurance file found.')
+
+        tabs.forget(tabs.select())
 
 
 def open_output(c_name, last_data):
@@ -234,7 +242,14 @@ def activate_tab(window, tab, act, start_btn, c_name, companies):
 
 
 def set_dir(window, c_name, btn, comp, last_data):
-    dir_name = askdirectory(title='Choose output folder', initialdir='.')
+    last_dir = Path(last_data[c_name]['output'])
+    if last_dir.exists():
+        initdir = last_dir
+    else:
+        initdir = Path.home()
+    
+    dir_name = askdirectory(title='Choose output folder', initialdir=initdir)
+
     if dir_name:
         window.nametowidget(comp + '.' + btn.winfo_name())["text"] = str(Path(dir_name).name)
         last_data[c_name]['output'] = str(Path(dir_name))
@@ -244,9 +259,16 @@ def set_dir(window, c_name, btn, comp, last_data):
 
 def set_datas(window, c_name, last_data, btn, comp, filetypes, key_name):
 
+    last_dir = Path(last_data[c_name][key_name]).parent
+    if last_dir.exists():
+        initdir = last_dir
+    else:
+        initdir = Path.home()
+        
+
     filename = askopenfilename(
         title='Open a file',
-        initialdir='.',
+        initialdir=initdir,
         filetypes=filetypes
     )
